\docType{data}
\name{clin.crc}
\alias{clin.crc}
\title{Colon cancer clinical dataset}
\format{A matrix with 10 rows and 2 columns}
\usage{
data(clin.crc)
}
\value{An example clinical dataset}
\description{
The example dataset consisits of 10 colon cancer patients, of which 5 had relapse (return of cancer to colon)
and the rest did not relapse. This example dataset is part of the complete dataset used in \cite{CRC}, and
can be accessed via G-DOC Plus at https://gdoc.georgetown.edu.
The column names are described below:
}
\details{
\itemize{
  \item Sample. Sample ID
  \item Label. Refers to the group label/outcome
  }
  More details on how this object was created is provided in the vignette titled "How to prepare Input data" in the CINdex package.

  }
\keyword{datasets}
