% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.heatmap.R
\name{comp.heatmap}
\alias{comp.heatmap}
\title{A comprehensive heatmap function that plots Chromosome and Cytoband heatmaps}
\usage{
comp.heatmap(R_or_C = "Regular", clinical.inf = NULL, genome.ucsc = NULL,
  in.folder.name = "output_chr_cin", out.folder.name = "output_chr_plots",
  plot.choice = "png", base.color = "black", thr.gain = c(2.5, 2.25, 2.1),
  thr.loss = c(1.5, 1.75, 1.9), V.def = 2:3, V.mode = c("sum", "amp",
  "del"))
}
\arguments{
\item{R_or_C}{The value'Regular' plots chromosome level heatmap and 'Cytobands' plots cytoband level
heatmaps}

\item{clinical.inf}{An n*2 matrix, the 1st column is 'sample name', the second is 'label'}

\item{genome.ucsc}{A Reference genome}

\item{in.folder.name}{Name of folder where the Chromsome CIN or Cytoband CIN objects are present}

\item{out.folder.name}{Name of folder where the Chromosome heatmaps or Cytoband heatmaps will be saved}

\item{plot.choice}{A choice of whether the heatmaps should be .png or .pdf format}

\item{base.color}{A choice of 'black' or 'white' base color for the heatmap (indicating no instability)}

\item{thr.gain}{A threshold above which will be set as gain}

\item{thr.loss}{A threshold below which will be set as loss}

\item{V.def}{There are 2 different CIN definitions - normalized (value=2) and un-normalized (value=3)}

\item{V.mode}{There are 3 options: 'sum', 'amp' and 'del'}
}
\value{
No value returned. If R_or_C='Regular', it will genearte chromosome level heatmap,
If R_or_C='Cytobands',it will generate cytoband level heatmap
}
\description{
When the \code{run.cin.chr} and \code{run.cyto.chr} functions are
called, we get Chromosome and Cytoband CIN values for various gain/loss threshold settings.
This \code{comp.heatmap} function can be used to pick the best threshold for the input data.
It plots heatmaps for two groups of interest (case and control) for all the input gain/loss threshold
settings. By visually checking the heatmaps, the user can pick the threshold/setting that shows the best
contrast between two groups of interest.
Steps:
#Step 1: Run cytoband CIN or chromosome CIN - using \code{run.cin.chr()} or \code{run.cin.cyto()}
#Step 2: Call this function to create chromosome or cytoband level heatmaps. Pick gain/loss threshold
appropriate for data.
See vignette for more details.
}
\examples{
###### Example 1 - Chromosome level

## Step 1: Run chromosome CIN
# This is how command should be run:
\dontrun{
run.cin.chr(grl.seg = grl.data)
}
# For this example, we run chr CIN on one threshold only
data("grl.data")
run.cin.chr(grl.seg = grl.data, thr.gain=2.25, thr.loss=1.75, V.def=3, V.mode="sum")

## Step 2: Plot chromosome level heatmap
# This is how the command must be called:
\dontrun{
comp.heatmap(R_or_C="Regular", clinical.inf=clin.crc, genome.ucsc=hg18.ucsctrack, thr.gain = 2.25,
thr.loss = 1.75,V.def = 3,V.mode = "sum")
}
# For this example, we run chr heatmap on one threshold only
comp.heatmap(R_or_C='Regular', clinical.inf=clin.crc, genome.ucsc=hg18.ucsctrack, thr.gain = 2.25,
thr.loss = 1.75,V.def = 3,V.mode = "sum")

###### Example 2 - Cytoband level

## Step 1 : Run cytoband CIN
# This is how command should be run:
\dontrun{
run.cin.cyto(grl.seg = grl.data,cnvgr=cnvgr.18.auto, snpgr=snpgr.18.auto,
genome.ucsc = hg18.ucsctrack)

## Step 2: Plot cytoband level heatmap

comp.heatmap(R_or_C="Cytobands", clinical.inf=clin.crc, genome.ucsc=hg18.ucsctrack,
thr.gain=2.25, thr.loss=1.75,V.def=3,V.mode="sum")
}

}
\seealso{
See accompanying vignette for end-to-end tutorial
}

