\docType{data}
\name{cyto.cin4heatmap}
\alias{cyto.cin4heatmap}
\title{Cytoband CIN T-test output}
\format{List}
\usage{
data(cyto.cin4heatmap)
}
\value{Cytoband CIN T-test output}
\description{
Example output obtained from running the T-test on Cytoband CIN object.
See accompanying vignette in the CINdex package for a complete tutorial
}
\keyword{datasets}
