\docType{data}
\name{cytobands.cin}
\alias{cytobands.cin}
\title{Cytoband CIN dataset}
\format{List}
\usage{
data(cytobands.cin)
}
\value{An example cytoband CIN}
\description{
Example output obtained from running the Cytoband CIN function in the CINdex package. Indicates chromsome instability index value for every cytoband.
}
\keyword{datasets}
