% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.genes.in.cyto.regions.R
\name{extract.genes.in.cyto.regions}
\alias{extract.genes.in.cyto.regions}
\title{Given an input of cytobands, it outputs a list of genes that are present in the cytoband regions}
\usage{
extract.genes.in.cyto.regions(cyto.cin4heatmapObj = NULL,
  genome.ucsc = NULL, gene.annotations = NULL,
  folder.name = "output_genename")
}
\arguments{
\item{cyto.cin4heatmapObj}{Output of the cytoband T test results}

\item{genome.ucsc}{Reference sequence}

\item{gene.annotations}{Information about CDS start and end positions, Gene names}

\item{folder.name}{Name of output folder}
}
\value{
Output files: The genes names present in the cytoband regions
}
\description{
Once the user has a list of cytobands of interest, one downstream application could
be to find the list of genes present in the cytoband regions. This \code{extract.genes.in.cyto.regions}
function can be used for this purpose. The following steps should be run before this function can
be called:
#Step 1 : Run cytoband CIN - using \code{run.cin.chr()}
#Step 2: Plot cytoband level heatmap - using \code{comp.heatmap()}
#Step 3: Go through heatmaps as select one appropriate threshold. Load the file.
#Step 4: Perform T test to find differentially expressed cytobands - using \code{ttest.cyto.cin.heatmap()}
#Step 5: Call this funtion to extract genes located in cytoband regions
#More details and tutorial are given in the accompanying vignette
}
\examples{
#For this example, we load example T test output object
data("cyto.cin4heatmap")
data("hg18.ucsctrack") #load Hg 18 reference annotation file
data("geneAnno") #load Gene annotations file
extract.genes.in.cyto.regions(cyto.cin4heatmapObj =cyto.cin4heatmap,
genome.ucsc = hg18.ucsctrack, gene.annotations = geneAnno)
}
\seealso{
See accompanying vignette for an end-to-end tutorial
}

