\docType{data}
\name{geneAnno}
\alias{geneAnno}
\title{CDS gene annotation file}
\format{A matrix}
\usage{
data(geneAnno)
}
\value{An example CDS gene annotation file}

\description{
A CDS gene annotation file with the following column names (obtained for human reference)
 \itemize{
  \item chrom. Chromosome number
  \item strand. Positive or negative strand
  \item cdsStart. CDS Start position
  \item cdsEnd. CDS end position
  \item GeneID. Gene symbol
  }
  More details on how this object was created is provided in the vignette titled "How to prepare Input data" in the CINdex package.
}
\keyword{datasets}
