\docType{data}
\name{hg18.ucsctrack}
\alias{hg18.ucsctrack}
\title{Human reference annotation file}
\format{GRanges object}
\usage{
data(hg18.ucsctrack)
}
\value{An example hg18 annotation file}

\description{
The reference annotation file used in the CIN algorithm. The example file used here is for Human Species hg18 and includes information about chromosome number, start and end position, name of cytoband and stain.

More details on how this object was created is provided in the vignette titled "How to prepare Input data" in the CINdex package.

}
\keyword{datasets}
