\docType{data}
\name{snpgr.18.auto}
\alias{snpgr.18.auto}
\title{Probe annotation file for Affymetrix Genome Wide Human SNP Array 6.0}
\format{A GRanges object}
\usage{
data(snpgr.18.auto)
}
\value{An example probe annotation file}
\description{
This is a probe annotation file for Affymetrix Genome Wide Human SNP Array 6.0. It contains annotation for only the SNP probes in this array and corresponds to hg18 reference genome.

The GRanges object contains details about probe name, chromosome number, physical location and strand. The annotation has been filtered to include only those probes that are located in autosomes.

More details on how this object was created is provided in the vignette titled "How to prepare Input data" in the CINdex package.

}
\keyword{datasets}
