% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.cyto.cin.heatmap.R
\name{ttest.cyto.cin.heatmap}
\alias{ttest.cyto.cin.heatmap}
\title{Performs T test on cytoband level CIN data, and plots heatmap}
\usage{
ttest.cyto.cin.heatmap(cytobands.cin.obj, clinical.inf, genome.ucsc,
  file.ext = "gainT_lossT_unnorm", folder.name = "output_ttest",
  combine.cyto.flag = FALSE)
}
\arguments{
\item{cytobands.cin.obj}{(eg. cytobands.cin_2.25_1.75_unnormalized_amp.Rdata), a list in which each cell is chromosome cin matrix}

\item{clinical.inf}{In a clinical.inf.Rdata is a two columns array, the 1st column is samplename, the 2nd is the label}

\item{genome.ucsc}{Reference sequence}

\item{file.ext}{Provide a meaningful file name extension. Ideally include the gain, loss threshold settings}

\item{folder.name}{Name of folder where the output files will be generated}

\item{combine.cyto.flag}{Whether or not to save the combine cytobands as a uni array rather than a list}
}
\value{
#Outputs:
     1. cyto.cin.uni.file.ext.Rdata (eg. cyto.cin.uni.gainT_lossT_unnormalized.Rdata)
     2. Heatmaps: eg. CIN relapse-free VS relapse for gainT_lossT_unnormalized_dendrogram.pdf
     3. Raw CIN array for the corresponding heatmap:
     #ttest.cyto.cin4heatmap.gainT_lossT_unnormalized.csv
     #ttest.cyto.cin4heatmap.gainT_lossT_unnormalized.Rdata
     4. T test results for all cytobands on the whole genome
     #ttest.cytobands.cin.gainT_lossT_unnormalized.txt
}
\description{
\code{ttest.cyto.cin.heatmap} to perform T test to find differentially expressed cytobands.
It also plots a heatmap after performing heirarchical clustering. When to use this function:
#Step 1: Run cytoband CIN - using \code{run.cin.chr()}.
#Step 2: Plot cytoband level heatmap - using \code{comp.heatmap()}.
#Step 3: Go through heatmaps as select one appropriate threshold. Load the file.
#Step 4: Call this function.
More details and tutorial are given in the accompanying vignette
}
\examples{
#For this example, we load an example cytoband CIN data
data("cytobands.cin")
data("clin.crc") # sample names with group information
data("hg18.ucsctrack") #hg18 reference file
ttest.cyto.cin.heatmap(cytobands.cin.obj = cytobands.cin,
clinical.inf = clin.crc, genome.ucsc = hg18.ucsctrack)
}
\seealso{
See accompaying vignette for a detailed end to end workflow tutorial
}

